+++ apps/control-center/src/app/shared/components/thrift-api-crud/domain/domain-thrift-editor/domain-thrift-editor.component.html
