+++ apps/control-center/src/app/shared/components/thrift-api-crud/domain/domain-thrift-editor/domain-thrift-editor.component.ts
@@ -13,8 +13,8 @@ import { DomainMetadataFormExtensionsService } from '../../../../services';
 import { BaseThriftFormSuperclass } from '../../thrift-forms/utils/thrift-form-superclass';
 
 @Component({
-    selector: 'cc-domain-thrift-form',
-    templateUrl: './domain-thrift-form.component.html',
+    selector: 'cc-domain-thrift-editor',
+    templateUrl: './domain-thrift-editor.component.html',
     providers: createControlProviders(() => DomainThriftFormComponent),
     imports: [CommonModule, ReactiveFormsModule, ThriftFormModule, ThriftEditorModule],
 })
