+++ apps/control-center/src/app/shared/components/thrift-api-crud/domain/domain-thrift-editor/index.ts
@@ -1,0 +1,1 @@
+export * from './domain-thrift-editor.component';
