--- apps/control-center/src/app/shared/components/thrift-api-crud/domain/domain-thrift-form/index.ts
@@ -1,1 +1,0 @@
-export * from './domain-thrift-form.component';
