+++ apps/control-center/src/app/shared/components/thrift-api-crud/domain/index.ts
@@ -1,5 +1,5 @@
 export * from './domain-thrift-viewer';
 export * from './domain-object-field';
 export * from './domain-thrift-form-dialog';
-export * from './domain-thrift-form';
+export * from './domain-thrift-editor';
 export * from './utils';
