+++ apps/control-center/src/app/shared/components/thrift-api-crud/domain2/create-domain-object-dialog/create-domain-object-dialog.component.html
@@ -10,11 +10,11 @@
             type="InsertOp"
         ></cc-domain-thrift-viewer>
     } @else {
-        <cc-domain-thrift-form
+        <cc-domain-thrift-editor
             [formControl]="control"
             namespace="domain_config_v2"
             type="InsertOp"
-        ></cc-domain-thrift-form>
+        ></cc-domain-thrift-editor>
     }
 
     <v-dialog-actions>
