+++ apps/control-center/src/app/shared/components/thrift-api-crud/domain2/create-domain-object-dialog/create-domain-object-dialog.component.ts
@@ -22,7 +22,7 @@ import { ValuesType } from 'utility-types';
 import { DomainService } from '../../../../../api/domain-config';
 import { APP_ROUTES } from '../../../../../app-routes';
 import { NavigateService } from '../../../../services';
-import { DomainThriftFormComponent } from '../../domain/domain-thrift-form';
+import { DomainThriftFormComponent } from '../../domain/domain-thrift-editor';
 import { DomainThriftViewerComponent } from '../../domain/domain-thrift-viewer';
 
 @Component({
