+++ apps/control-center/src/app/shared/components/thrift-api-crud/domain2/edit-domain-object-dialog/edit-domain-object-dialog.component.ts
@@ -38,7 +38,7 @@ import {
 } from '../../../../../api/domain-config';
 import { APP_ROUTES } from '../../../../../app-routes';
 import { NavigateService } from '../../../../services';
-import { DomainThriftFormComponent } from '../../domain/domain-thrift-form';
+import { DomainThriftFormComponent } from '../../domain/domain-thrift-editor';
 import { DomainThriftViewerComponent } from '../../domain/domain-thrift-viewer';
 
 enum Step {
